Rename chat bot to HASSTNT and change the skeleton
