+++ src/main/java/HASSTNT.java
@@ -1,0 +1,13 @@
+public class HASSTNT{
+    private static final String welcome_message = "____________________________________________________________\n" +
+            "Hello! I'm HASSTNT\n" +
+            "What can I do for you?\n" +
+            "____________________________________________________________\n" +
+            " Bye. Hope to see you again soon!\n"+
+            "____________________________________________________________\n";
+
+    public static void main(String[] args) {
+        System.out.println(welcome_message);
+
+    }
+    }
